pyproject.toml
[tool.poetry]
name = "goutte"
version = "2.0.0"
description = "DigitalOcean snapshot automation"
readme = "README.md"
license = "GPL-3.0+"
homepage = "https://github.com/tomochain/goutte"
repository = "https://github.com/tomochain/goutte"
keywords = ["digitalocean", "do", "snapshot", "snapshots", "backup"]
authors = ["etienne-napoleone <etienne@tomochain.com>"]

[tool.poetry.dependencies]
python = "^3.6"
python-digitalocean = "^1.16"
colorlog = "^3.1"
click = "^7.1.2"
PyYAML = "^5.4.1"
yamale = "^3.0.4"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
flake8 = "^3.6"
coveralls = "^1.5"
mypy = "^0.800"
black = "^20.8b1"

[tool.poetry.scripts]
goutte = "goutte.main:entrypoint"
